fix character interfaces
