--- src/app/shared/models/characters/CharacterDataContainer.ts
@@ -1,9 +1,0 @@
-import { Character } from "./Character";
-
-export interface CharacterDataContainer{
-    offset:number;
-    limit:number;
-    total:number;
-    count:number;
-    results: [Character];
-}
