--- src/app/shared/models/characters/CharacterDataWrapper.ts
@@ -1,11 +1,0 @@
-import { CharacterDataContainer } from "./CharacterDataContainer";
-
-export interface CharacterDataWrapper{
-    code:number;
-    status:string;
-    copyright:string;
-    attributionText:string;
-    attributionHTML:string;
-    data:CharacterDataContainer;
-    etag:string;
-}
