--- src/app/shared/models/characters/ComicList.ts
@@ -1,8 +1,0 @@
-import { ComicSummary } from "./ComicSummary";
-
-export interface ComicList{
-    available:number;
-    returned:number;
-    collectionURI:string;
-    items:ComicSummary[];
-}
