--- src/app/shared/models/characters/ComicSummary.ts
@@ -1,4 +1,0 @@
-export interface ComicSummary{
-    resourceURI:string;
-    name:string;
-}
