--- src/app/shared/models/characters/SeriesList.ts
@@ -1,8 +1,0 @@
-import { SeriesSummary } from "./SeriesSummary";
-
-export interface SeriesList{
-    available:number;
-    returned:number;
-    collectionURI:string;
-    items:[SeriesSummary];
-}
