--- src/app/shared/models/characters/URLCharacters.ts
@@ -1,4 +1,0 @@
-export interface Url{
-    type:string;
-    url:string;
-}
