--- src/app/shared/models/characters/image.ts
@@ -1,4 +1,0 @@
-export interface Image{
-    path:string;
-    extension:string;
-}
